+++ dags/dwh/rating/comment.py
@@ -7,6 +7,8 @@ from airflow.datasets import Dataset
 from airflow.decorators import dag, task
 from airflow.providers.postgres.operators.postgres import PostgresOperator
 
+from plugins.features import get_sql_code
+
 with DAG(
     dag_id="DWH_RATING.comment",
     schedule=[Dataset("ODS_RATING.comment")],
@@ -23,6 +25,7 @@ with DAG(
         task_id="comment_hist",
         postgres_conn_id="postgres_dwh",
         sql="comment.sql",
+        doc_md=get_sql_code('comment.sql'),
         inlets=[Dataset("ODS_RATING.comment")],
         outlets=[Dataset("DWH_RATING.comment")],
     )
